+++ Content.Shared/Eye/Blinding/Components/BlindnessStatusEffectComponent.cs
@@ -6,4 +6,4 @@ namespace Content.Shared.Eye.Blinding.Components;
 /// Prevents the target from seeing while active.
 /// </summary>
 [NetworkedComponent, RegisterComponent]
-public sealed partial class TemporaryBlindnessStatusEffectComponent : Component;
+public sealed partial class BlindnessStatusEffectComponent : Component;
